Refactor device acceleration structure storage
We need to maintain a list of all currently active acceleration structures. This stems from a somewhat complicated issue. In shaders, vk uses addresses but metal uses indices tied to an array of acc structs passed in to the command. Since we have no efficient or guaranteed way of deducing what acc structs are used in a TLAS given the instance buffer, the best way is to maintain all instances and have their index into the global list tied to their respective address shared with the client.